+++ scripts/embedding/xsb/xsb_logtalk_xwam.sh
@@ -6,7 +6,7 @@
 ##   and runtime and optionally an application.xwam file with a Logtalk
 ##   application
 ##
-##   Last updated on January 30, 2025
+##   Last updated on March 23, 2025
 ##
 ##   This file is part of Logtalk <https://logtalk.org/>
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -28,7 +28,7 @@
 
 
 print_version() {
-	echo "$(basename "$0") 0.19"
+	echo "$(basename "$0") 0.20"
 	exit 0
 }
 
@@ -166,8 +166,8 @@ while getopts "cd:t:p:l:s:vh" option; do
 		s) s_arg="$OPTARG";;
 		l) l_arg="$OPTARG";;
 		v) print_version;;
-		h) usage_help; exit;;
-		*) usage_help; exit;;
+		h) usage_help; exit 0;;
+		*) usage_help; exit 1;;
 	esac
 done