+++ scripts/install.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ##
 ##   Logtalk installation script
-##   Last updated on March 22, 2025
+##   Last updated on March 23, 2025
 ##
 ##   This file is part of Logtalk <https://logtalk.org/>
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -42,7 +42,7 @@ default_directory=logtalk-$version
 
 print_version() {
 	echo "Current $(basename "$0") version:"
-	echo "  0.13"
+	echo "  0.14"
 	exit 0
 }
 
@@ -69,8 +69,8 @@ while getopts "p:d:vh" option; do
 		v) print_version;;
 		p) prefix_argument="$OPTARG";;
 		d) directory_argument="$OPTARG";;
-		h) usage_help; exit;;
-		*) usage_help; exit;;
+		h) usage_help; exit 0;;
+		*) usage_help; exit 1;;
 	esac
 done