+++ scripts/update_html_docs.sh
@@ -5,7 +5,7 @@
 ##   Logtalk script for updating the HTML core, library, tools, ports,
 ##   contributions, and (optionally) packs documentation
 ##
-##   Last updated on March 14, 2025
+##   Last updated on March 23, 2025
 ##
 ##   This file is part of Logtalk <https://logtalk.org/>
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -68,7 +68,7 @@ set_goal() {
 }
 
 print_version() {
-	echo "$(basename "$0") 0.26"
+	echo "$(basename "$0") 0.27"
 	exit 0
 }
 
@@ -98,8 +98,8 @@ while getopts "vip:m:d:h" option; do
 		v) print_version;;
 		p) p_arg="$OPTARG";;
 		i) include_packs='true';;
-		h) usage_help; exit;;
-		*) usage_help; exit;;
+		h) usage_help; exit 0;;
+		*) usage_help; exit 1;;
 	esac
 done