+++ tools/lgtdoc/xml/lgt2xml.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ##
 ##   XML documenting files to XML conversion script
-##   Last updated on March 21, 2025
+##   Last updated on March 23, 2025
 ##
 ##   This file is part of Logtalk <https://logtalk.org/>
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -25,7 +25,7 @@
 
 
 print_version() {
-	echo "$(basename "$0") 1.3"
+	echo "$(basename "$0") 1.4"
 	exit 0
 }
 
@@ -116,8 +116,8 @@ while getopts "vf:i:t:h" option; do
 		f) f_arg="$OPTARG";;
 		i) i_arg="$OPTARG";;
 		t) t_arg="$OPTARG";;
-		h) usage_help; exit;;
-		*) usage_help; exit;;
+		h) usage_help; exit 0;;
+		*) usage_help; exit 1;;
 	esac
 done